Fix generation of WSI WSDL with typekeys
Typekey types are no longer enums, but we want them to act like enums
for the purposes of WSI WSDL generation. To achieve this, allow for
custom MarshalInfo objects to be created per type. Register a custom
MarshalInfoFactory for all typekey types in PLXSDTypeProvider.

As part of this change, re-factored EnumMarshalInfo into a base class
with two subclasses: one for IEnumTypes that behaves as before, and
another for typekey types. Factored out an IEnumData interface that is
extended both by IEnumType and ITypeListTypeInternal (which is no longer
an IType), and that is used by the new EnumMarshalInfo base class to get
enum info.