Replace extended type proxies with generated code
In order to support runtime assignability of Java types representing
entities to IEntityType, we introduced this Java extensions mechanism
earlier in Ferrite. The original implementation wrapped the IJavaType and
the IEntityType data in a java.lang.reflect.Proxy, and delegated all
calls using reflection. It was set up such that all calls to IJavaType
methods went to the IJavaType instance, so that the IEntityType could
not override any behavior of IJavaType. The same is applied to
IJavaPropertyInfo and IEntityPropertyInfo.

However, performance testing revealed the use of java.lang.reflect.Proxy
and reflection to be a huge hotspot, as the type and property objects are used a lot,
particularly in ORM code. To fix this, we generate classes that do the
same thing, essentially implementing the proxy ourselves.

There are two different flavors of proxy: one for java types and one for
java properties. Each has a base class that implements the
java.lang.Object methods equals(), hashCode(), and toString(). All other
methods in the generated class simply delegate to their underlying
object.